+++ src/services/home/type.ts
@@ -1,0 +1,5 @@
+import { RecipeSimple } from '../recipe/type';
+
+export type HomeRecipe = RecipeSimple & {
+  user: { username: string };
+};
