+++ src/services/recipes/index.ts
@@ -2,7 +2,7 @@ import { fetchAPI } from '@/services/api';
 
 import { lightTrim } from '@/utils/normalize';
 
-import { GetRecipesRes, HomeRecipe } from './type';
+import { GetRecipesRes } from './type';
 import { RecipeDetail } from '../recipe/type';
 import { FetchResult } from '../type';
 import { createAsyncErrorMessage, withSafeAsync } from '../utils';
@@ -28,20 +28,6 @@ export const getRecipes = withSafeAsync(
     recommend: [],
   },
 );
-
-export const getHomeRecipes = withSafeAsync(async (): FetchResult<
-  HomeRecipe[]
-> => {
-  const response = await fetchAPI('/recipes/home', {
-    cache: 'force-cache',
-  });
-
-  if (response.ok) return { ok: true, data: response.data };
-
-  throw new Error(
-    createAsyncErrorMessage(response.res, `Failed to fetch home recipes`),
-  );
-}, []);
 
 export const createRecipe = async (
   data: FormData,
